requirements.txt
alembic==1.8.1
aniso8601==9.0.1
antiorm==1.2.1
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
db==0.1.1
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.5
py==1.11.0
pydantic==1.10.2
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
SQLAlchemy-serializer==1.4.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
